backend_merged/requirements.txt
# Ques Backend - Merged Requirements
# Combines dependencies from both backend_p12 and backend_p34
# Install: pip install -r requirements.txt

# ⚠️ EXTERNAL API DEPENDENCIES REQUIRED FOR PROJECT IDEAS ⭐ NEW
# - SearchAPI key (SEARCHAPI_KEY): https://www.searchapi.io/
# - DeepSeek API key (DEEPSEEK_API_KEY_AGENT): https://api.deepseek.com/

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Production WSGI server
gunicorn==21.2.0

# Data validation and configuration  
pydantic[email]>=2.9.0
email-validator>=2.0.0

# HTTP requests for external APIs
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.0

# Web scraping and content extraction (Project Idea Agent) ⭐ NEW
crawl4ai==0.2.77
beautifulsoup4==4.12.2

# Environment variables
python-dotenv==1.0.0

# PostgreSQL database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.25
alembic>=1.12.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File uploads and multipart
python-multipart>=0.0.6

# Database migrations
alembic==1.13.0

# JWT and Authentication
python-jose[cryptography]==3.3.0
pyjwt>=2.8.0

# Password hashing and security
passlib[bcrypt]==1.7.4
cryptography>=41.0.0

# Email validation
email-validator==2.0.0

# Enhanced security and monitoring
structlog>=23.0.0

# Redis for caching and sessions
redis>=5.0.0
hiredis>=2.3.0

# Vector embeddings and AI
sentence-transformers>=2.7.0
huggingface_hub>=0.19.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# Tencent Cloud Services
tcvectordb
tencentcloud-sdk-python>=3.0.1000

# DeepSeek API for tag extraction and project idea generation ⭐ NEW
openai>=1.0.0  # Compatible API client for DeepSeek API

# Project Idea Agent Dependencies ⭐ NEW
# Web search and content extraction
requests==2.31.0
httpx==0.25.0
aiohttp==3.8.6

# Web scraping (required for project idea agent)
crawl4ai==0.2.77
beautifulsoup4==4.12.2
jsonschema>=4.0.0

# WebSocket support for real-time features
websockets>=11.0.0

# Rate limiting
slowapi>=0.1.9

# Image processing (for profile pictures and project images)
Pillow>=10.0.0

# File type detection
python-magic>=0.4.27

# Background tasks
celery>=5.3.0

# Additional utilities
python-multipart>=0.0.6  # For file uploads
httpx>=0.25.0  # Modern HTTP client
pytest>=7.4.0  # Testing framework
pytest-asyncio>=0.21.0  # Async testing

# Development dependencies (optional)
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking
